requirements.txt
google-generativeai
pandas
numpy
python-dotenv
matplotlib
seaborn
scipy 
tqdm